Open and close performance metrics
PressB script tries to open and close the performance metrics with the button B on the controller. However, the metrics overlay overrides this with its own method. The overly key bindings to show and hide the metrics only supports keys on a keyboard.

Testing the apk on the headset with a bluetooth keyboard properly shows and hides the performance metrics. Unless there is a way to choose a key on the controller to do this, or create an option in the pause menu, the only way to show or hide the overlay will be with a keyboard.