Sprint 6 Tasks
In this push contains the integration of all glassware into the lab. Any container can be used to complete the liquids lab.

I added the ability to reset the lab in the pause menu, it simply reloads the scene.

Each time the lab it performed, the liquids needed are randomized, providing the user with a variety of possible outcomes.

I added the ability for the user to request more feedback from the AI after completing the lab.

I moved the AI feedback to the wall due to motion sickness issues when trying to read feedback attached to Rig.